requirements_auth.txt
# Authentication and Database Requirements
# Add these to your existing requirements.txt

Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
python-dotenv==1.0.0
psycopg2-binary==2.9.9  # PostgreSQL adapter for production
Werkzeug==3.0.1  # For password hashing
